update(NewShow): Cleaned up the code, removed and renamed some files. Details in PR
